+++ app/studio/models.py
@@ -1,0 +1,3 @@
+"""Studio 앱은 관리자 워크플로 전용 뷰를 제공하며, 현재 별도 모델을 사용하지 않습니다."""
+
+# 모델이 필요해지면 이 파일에 정의합니다.
