+++ app/user/urls.py
@@ -2,8 +2,6 @@
 from django.urls import path
 
 from .views import (
-    AdminDashboardView,
-    AdminSubmissionDetailView,
     CustomLoginView,
     LogoutRedirectView,
     ProfileView,
@@ -18,12 +16,6 @@ urlpatterns = [
     path("logout/", LogoutRedirectView.as_view(), name="logout"),
     path("profile/", ProfileView.as_view(), name="profile"),
     path("signup/", SignupView.as_view(), name="signup"),
-    path("admin/dashboard/", AdminDashboardView.as_view(), name="admin_dashboard"),
-    path(
-        "admin/submissions/<int:pk>/",
-        AdminSubmissionDetailView.as_view(),
-        name="admin_submission_detail",
-    ),
     path(
         "submissions/<int:pk>/edit/",
         SubmissionUpdateView.as_view(),
