+++ app/user/views.py
@@ -6,10 +6,8 @@ from django.contrib.auth.views import LoginView
 from django.http import HttpResponseRedirect
 from django.shortcuts import redirect
 from django.urls import reverse, reverse_lazy
-from django.utils import timezone
 from django.views import View, generic
 
-from app.post.models import Post
 from app.submission.forms import SubmissionForm
 from app.submission.models import Submission, SubmissionStatus
 
@@ -26,8 +24,8 @@ class CustomLoginView(LoginView):
         redirect_to = self.get_redirect_url()
         if redirect_to:
             return redirect_to
-        if self.request.user.is_superuser:
-            return reverse("user:admin_dashboard")
+        if self.request.user.is_staff:
+            return reverse("studio:dashboard")
         return reverse("post:list")
 
 
@@ -58,44 +56,6 @@ class SignupView(generic.FormView):
         return super().form_valid(form)
 
 
-class AdminDashboardView(LoginRequiredMixin, UserPassesTestMixin, generic.TemplateView):
-    """슈퍼유저 전용 대시보드 화면입니다."""
-
-    template_name = "user/admin_dashboard.html"
-
-    def test_func(self):
-        """요청 사용자가 슈퍼유저인지 확인합니다."""
-        return self.request.user.is_superuser
-
-    def handle_no_permission(self):
-        """권한이 없으면 게시판으로, 비로그인이라면 기본 동작을 따릅니다."""
-        if self.request.user.is_authenticated:
-            return redirect("post:list")
-        return super().handle_no_permission()
-
-    def get_context_data(self, **kwargs):
-        """최근 신청과 게시글 요약 정보를 채워 넣습니다."""
-        context = super().get_context_data(**kwargs)
-        context["pending_submissions"] = (
-            Submission.objects.filter(status__in=[SubmissionStatus.SUBMITTED, SubmissionStatus.IN_REVIEW])
-            .select_related("bike", "bike__spec")
-            .order_by("-created_at")[:5]
-        )
-        context["in_progress_submissions"] = (
-            Submission.objects.filter(status=SubmissionStatus.IN_PROGRESS)
-            .select_related("bike", "bike__spec")
-            .order_by("-reviewed_at", "-created_at")[:5]
-        )
-        context["recent_posts"] = Post.objects.order_by("-created_at")[:5]
-        context["total_pending"] = Submission.objects.filter(
-            status__in=[SubmissionStatus.SUBMITTED, SubmissionStatus.IN_REVIEW]
-        ).count()
-        context["total_in_progress"] = Submission.objects.filter(
-            status=SubmissionStatus.IN_PROGRESS
-        ).count()
-        return context
-
-
 class LogoutRedirectView(LoginRequiredMixin, View):
     """로그아웃하고 홈으로 돌려보내는 뷰입니다."""
 
@@ -103,55 +63,6 @@ class LogoutRedirectView(LoginRequiredMixin, View):
         """요청 방식을 가리지 않고 로그아웃한 뒤 홈으로 보냅니다."""
         logout(request)
         return redirect("post:list")
-
-
-class AdminSubmissionDetailView(LoginRequiredMixin, UserPassesTestMixin, generic.DetailView):
-    """신청서를 열람하고 상태를 바꾸는 관리자 뷰입니다."""
-
-    model = Submission
-    template_name = "user/admin_submission_detail.html"
-    context_object_name = "submission"
-
-    def test_func(self):
-        """요청 사용자가 슈퍼유저인지 확인합니다."""
-        return self.request.user.is_superuser
-
-    def get(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        submission = self.object
-        if submission.status == SubmissionStatus.SUBMITTED:
-            submission.status = SubmissionStatus.IN_REVIEW
-            submission.reviewer = request.user
-            submission.reviewed_at = timezone.now()
-            submission.save(update_fields=["status", "reviewer", "reviewed_at"])
-        context = self.get_context_data(object=submission)
-        return self.render_to_response(context)
-
-    def post(self, request, *args, **kwargs):
-        submission = self.get_object()
-        action = request.POST.get("action")
-        if action == "start_draft":
-            submission.status = SubmissionStatus.IN_PROGRESS
-            submission.reviewer = request.user
-            submission.reviewed_at = submission.reviewed_at or timezone.now()
-            submission.save(update_fields=["status", "reviewer", "reviewed_at"])
-            messages.success(request, "포스팅 준비를 시작했습니다. 자동 임시저장이 활성화됩니다.")
-            if submission.result_post:
-                edit_url = f"{reverse('post:edit', kwargs={'slug': submission.result_post.slug})}?submission={submission.pk}"
-                return redirect(edit_url)
-            create_url = f"{reverse('post:create')}?submission={submission.pk}"
-            return redirect(create_url)
-        if action == "reject":
-            reason = (request.POST.get("rejection_reason") or "").strip()
-            if not reason:
-                messages.error(request, "반려 사유를 입력해 주세요.")
-            else:
-                submission.status = SubmissionStatus.REJECTED
-                submission.rejection_reason = reason
-                submission.result_post = None
-                submission.save(update_fields=["status", "rejection_reason", "result_post"])
-                messages.success(request, "신청서를 반려했습니다.")
-        return redirect(self.request.path)
 
 
 class SubmissionUpdateView(LoginRequiredMixin, UserPassesTestMixin, generic.UpdateView):
