+++ config/urls.py
@@ -24,6 +24,7 @@ urlpatterns = [
     path("", include("app.post.urls")),
     path("users/", include("app.user.urls")),
     path("ckeditor/", include("ckeditor_uploader.urls")),
+    path("studio/", include("app.studio.urls")),
     path("api/", include(router.urls)),
     path("api/schema/", SpectacularAPIView.as_view(), name="api-schema"),
     path("api/docs/", SpectacularSwaggerView.as_view(url_name="api-schema"), name="api-docs"),
