+++ templates/studio/dashboard.html
@@ -1,10 +1,10 @@
 {% extends "base.html" %}
 
-{% block title %}관리자 대시보드{% endblock %}
+{% block title %}Studio 대시보드{% endblock %}
 
 {% block content %}
   <section>
-    <h2>관리자 대시보드</h2>
+    <h2>Studio 대시보드</h2>
     <p>{{ request.user.username }}님, 환영합니다. 아래에서 주요 관리 작업을 진행하세요.</p>
     <p>
       <a href="{% url 'post:create' %}" role="button" class="contrast">새 게시글 작성하기</a>
@@ -44,7 +44,7 @@
               </td>
               <td>{{ submission.created_at|date:"Y-m-d H:i" }}</td>
               <td>
-                <a href="{% url 'user:admin_submission_detail' submission.id %}">검토하기</a>
+                <a href="{% url 'studio:submission_detail' submission.id %}">검토하기</a>
                 {% if submission.result_post %}
                   · <a href="{% url 'post:edit' submission.result_post.slug %}?submission={{ submission.id }}">수정</a>
                 {% endif %}
@@ -78,7 +78,7 @@
               {% endif %}
             </strong>
             · {{ submission.created_at|date:"Y-m-d" }}
-            <span> — <a href="{% url 'user:admin_submission_detail' submission.id %}">세부 내용 확인</a>
+            <span> — <a href="{% url 'studio:submission_detail' submission.id %}">세부 내용 확인</a>
               {% if submission.result_post %}
                 | <a href="{% url 'post:edit' submission.result_post.slug %}?submission={{ submission.id }}">게시글 수정</a>
               {% endif %}
@@ -118,7 +118,7 @@
     <ul>
       <li><a href="{% url 'admin:index' %}">Django 기본 관리자 이동</a></li>
       <li><a href="{% url 'post:list' %}">게시글 목록 보기</a></li>
-      <li><a href="{% url 'user:profile' %}">프로필 관리</a></li>
+      <li><a href="{% url 'studio:dashboard' %}">Studio 홈</a></li>
     </ul>
   </section>
 {% endblock %}
