+++ templates/studio/submission_detail.html
@@ -4,7 +4,7 @@
 
 {% block content %}
   <section>
-    <a href="{% url 'user:admin_dashboard' %}">&larr; 대시보드로 돌아가기</a>
+    <a href="{% url 'studio:dashboard' %}">&larr; 대시보드로 돌아가기</a>
     <h2>소개 신청 상세</h2>
     <p><strong>제목:</strong> {{ submission.title|default:"(제목 없음)" }}</p>
     <dl>
